+++ src/helpers/Cp.php
@@ -147,7 +147,7 @@ class Cp
             if (!empty($licenseAlerts)) {
                 if ($canSettleUp) {
                     if ($path !== 'plugin-store/buy-all-trials') {
-                        $alerts[] = Craft::t('app', 'There are license trials that require payment.') . ' ' .
+                        $alerts[] = Craft::t('app', 'There are trial licenses that require payment.') . ' ' .
                             Html::a(Craft::t('app', 'Buy now'), UrlHelper::cpUrl('plugin-store/buy-all-trials'), ['class' => 'go']);
                     }
                 } else {
@@ -230,9 +230,10 @@ class Cp
      * @param string $context The context the element is going to be shown in (`index`, `field`, etc.)
      * @param string $size The size of the element (`small` or `large`)
      * @param string|null $inputName The `name` attribute that should be set on the hidden input, if `$context` is set to `field`
-     * @param bool $showStatus Whether the elemnet status should be shown (if the element type has statuses)
+     * @param bool $showStatus Whether the element status should be shown (if the element type has statuses)
      * @param bool $showThumb Whether the element thumb should be shown (if the element has one)
      * @param bool $showLabel Whether the element label should be shown
+     * @param bool $showDraftBadge Whether to show the “Draft” badge beside the label if the element is a draft
      * @return string
      * @since 3.5.8
      */
@@ -243,7 +244,8 @@ class Cp
         ?string $inputName = null,
         bool $showStatus = true,
         bool $showThumb = true,
-        bool $showLabel = true
+        bool $showLabel = true,
+        bool $showDraftBadge = true
     ): string {
         $label = $element->getUiLabel();
 
@@ -302,7 +304,7 @@ class Cp
             $htmlAttributes['class'] .= ' error';
         }
 
-        if ($element::hasStatuses()) {
+        if ($showStatus && $element::hasStatuses()) {
             $htmlAttributes['class'] .= ' hasstatus';
         }
 
@@ -376,10 +378,59 @@ class Cp
                 $html .= $encodedLabel;
             }
 
+            if ($showDraftBadge && $element->getIsDraft()) {
+                $html .= Html::tag('span', Craft::t('app', 'Draft'), [
+                    'class' => 'draft-label',
+                ]);
+            }
+
             $html .= '</span></div>';
         }
 
         $html .= '</div>';
+
+        return $html;
+    }
+
+    /**
+     * Returns element preview HTML, for a list of elements.
+     *
+     * @param ElementInterface[] $elements The elements
+     * @param string $size The size of the element (`small` or `large`)
+     * @param bool $showStatus Whether the element status should be shown (if the element type has statuses)
+     * @param bool $showThumb Whether the element thumb should be shown (if the element has one)
+     * @param bool $showLabel Whether the element label should be shown
+     * @param bool $showDraftBadge Whether to show the “Draft” badge beside the label if the element is a draft
+     * @return string
+     * @since 3.6.3
+     */
+    public static function elementPreviewHtml(
+        array $elements,
+        string $size = self::ELEMENT_SIZE_SMALL,
+        bool $showStatus = true,
+        bool $showThumb = true,
+        bool $showLabel = true,
+        bool $showDraftBadge = true
+    ): string {
+        if (empty($elements)) {
+            return '';
+        }
+
+        $first = array_shift($elements);
+        $html = static::elementHtml($first, 'index', $size, null, $showStatus, $showThumb, $showLabel, $showDraftBadge);
+
+        if (!empty($elements)) {
+            $otherHtml = '';
+            foreach ($elements as $other) {
+                $otherHtml .= static::elementHtml($other, 'index', $size, null, $showStatus, $showThumb, $showLabel, $showDraftBadge);
+            }
+            $html .= Html::tag('span', '+' . Craft::$app->getFormatter()->asInteger(count($elements)), [
+                'title' => implode(', ', ArrayHelper::getColumn($elements, 'title')),
+                'class' => 'btn small',
+                'role' => 'button',
+                'onclick' => 'jQuery(this).replaceWith(' . Json::encode($otherHtml) . ')',
+            ]);
+        }
 
         return $html;
     }